local weather bot v1.3 add chose city function. add schedule forecast update. add post construct data generation. add batches. fix minor bugs. add logging.

+++ src/main/java/com/example/vanotificator/repository/TelegramUserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TelegramUserRepository extends JpaRepository<TelegramUser, Integer> {
+public interface TelegramUserRepository extends JpaRepository<TelegramUser, Long> {
 
     boolean existsByChatId(long chatId);
